kernel: security audit fixes — C1-C6, H3-H4, M5
Critical fixes:
- C1: procfs UAF — add g_pid_cmdline pool, stop overwriting g_pid_status
- C2: heap corruption — kill process + schedule() instead of infinite loop
- C3: ext2 consistency — write inode after each i_blocks increment
- C4: shm UAF — reject shm_at on IPC_RMID'd segments (-EIDRM),
  skip marked_rm segments in shm_get lookup
- C6: ELF W^X — parse p_flags for segment permissions, re-protect
  after relocations, only re-protect full pages within non-writable
  segments (partial pages may be shared with .data)

High fixes:
- H3: execve_copy_user_str — add upfront user_range_ok check
- H4: rq_remove_if_queued — scan all priority queues in both
  active and expired runqueues (not just current priority)

Low fixes:
- M5: frame_refcount uint16_t → uint32_t to prevent overflow

Also adds EIDRM/ENOMSG errno definitions.

Tests: 120/120 QEMU, 33/33 battery, 69/69 host — zero regressions

+++ include/pmm.h
@@ -46,8 +46,8 @@ void pmm_free_page(void* ptr);
 
 // Reference counting for Copy-on-Write
 void pmm_incref(uintptr_t paddr);
-uint16_t pmm_decref(uintptr_t paddr);
-uint16_t pmm_get_refcount(uintptr_t paddr);
+uint32_t pmm_decref(uintptr_t paddr);
+uint32_t pmm_get_refcount(uintptr_t paddr);
 
 // Helper to print memory stats
 void pmm_print_stats(void);
